+++ plugins/version.client.ts
@@ -1,0 +1,4 @@
+export default defineNuxtPlugin(() => {
+  const { version, title } = useRuntimeConfig().public;
+  console.log(`${title} v${version}`);
+});